Dataset (#4)
* feat: Module datasets with functional Dataset

A new struct to handle input features with Zarr to reduce the memory
footprint. The cost is calculated on demand and kept in a new Zarr
container with a in memory cache for faster access.

* test: Removing dataset tests from core lib

The Dataset was moved to module dataset, and its tests with it.

* Ignore edges for now

Once it reaches the edges, stop early and warn. Let's postpone that for
now.

* log: Moving dbg to traditional logs

* fix: j is the second element in the tuple!

* test: Sample creates a Zarr dataset with multiple variables

* cfg: Adding nrel-transmission as CLI dependency

* Activating module dataset

* log: Adding warn level logs

* style:

* feat: Minimalist CLI to support Dataset development

* clean: Removing duplicate code already in module dataset

* clean: Removing root dataset in favor of dataset module

* feat: Tracking already calculated

Use a 2D bool index to track chunks already calculated.

* style:

* log:

* log: Debug when calculating a new chunk

This is critical to monitor if we are not running it more than a single
time. It deserves a debug level.

* refact: Defining cost storate from /A storage

Reproduce shape and chunk shape from /A.

* test: Diverting core tests to use samples from dataset

* refact: Allowing dataset to be used by root level

pub(super)

* clean:

* test: Making sample more flexible on the size

* fix: It's now required to inform cache size

* feat: resolve()

Prototype of a public interface function that minimizes the number of
public resources from core library.

* style:

* cfg: Updating Cargo.lock

* cfg: Bump versions

* style:

* fix: It doesn't need to be mutable

* cfg: Bump version

* style:

* style: