fix: Avoid repeated chunk calculation
A logical error using a RwLock would result a high probability of
redundant calculation and recording of cost, affecting the performance of
concurrent jobs.

I was not re-checking the index after the lock was acquired, thus
concurrent processes on the same target could end calculating the cost
for a chunk already recorded. Now, once the lock is finally acquired, it
double checks if it is that chunk is indeed empty before finally
calculating it.