LazyChunk (#80)
* feat: Chunk

* refact: Using LazyChunk instead of passing a Zarr object

* importing LazyChunk

* style

* refact: Renaming to features

* refact: Simplify for now and use chunk indices as u32

We certainly won't need u64 for this, but let's keep everything
consistent on u64 for now and reduce that when it's time to optimize.

* doc, fix: Updating documentation to reflect changes

* fix: Renamed to ci/cj

* refact: Renamed to calculate instead of calculate_chunk

The `calculate` is now agnostic on the domain. It just respond to the
given collection subset. Whoever calls this function is in charge of
defining the target domain.

* fix: i/j were renamed to ci/cj

* feat: Providing ci/cj from LazyChunk

Reinstating log info on the chunk indices.

* Ideas and thoughts

* Not used/verified yet

* clean: Unecessary type conversion

* style: