fix: Correct cost transition
The cost value read is the average value for the grid point. Before we
were simplifying and just taking the next grid point value as the
transition.

The correct estimate is half the path is along the origin gridpoint,
plus half of the target grid point to reach the center of the target
gridoint, therefore, the transition wheight should be the average
between the current grid point cost and the target gridpoint cost.

Note that the same principle is valid for diagonals, it is still the
average of both values, but we have to scale for the longer distance
along the diagonal, thus a sqrt(2) factor.