Cached data subset (#121)
* feat: LazySubset

* Load LazySubset in module dataset

* refact: Renaming cost dataset to swap

Help to distinguish between the variable 'cost' and the swap dataset
formerly also called 'cost'.

* fix: Missing convert a few places from cost to swap

* doc: lazy_subset

* refact: Calculate cost chunk with LazySubset

Instead of assuming that cost chunk shape is equal to all source
variables' chunks, we now request input features with ArraySubset.

* style: Allow deadcode for ci/cj

Those will be removed, but for now let's just silence the warning.

* LazySubset now returns Result<>

It should be considered an error to request a features variable that
doesn't exist or a subset outside the domain.

* fix: Variable cost is stored in swap, not source