cfg: Trigger benches on some PRs
The Zarrs upgrade to 0.22 was an example of broken migration missed by
the benchmark procedure. It was specifically related on the creation of
test cases.

The benchmark workflow is a heavy one, so we shall minimize the use of
it, but let's re-activate it on `pull_request` if there are changes on
the Rust side.