Fix benchmark for new zarrs-0.22 syntax (#151)
* cfg: Trigger benches on some PRs

The Zarrs upgrade to 0.22 was an example of broken migration missed by
the benchmark procedure. It was specifically related on the creation of
test cases.

The benchmark workflow is a heavy one, so we shall minimize the use of
it, but let's re-activate it on `pull_request` if there are changes on
the Rust side.

* fix: Conforming bench with new (Zarrs) ArrayBuilder's syntax

* cfg: Save benchmark only when push to main

Otherwise, it would save all the intermediate states during a PR
development.